% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust_infection_to_report}
\alias{adjust_infection_to_report}
\title{Adjust from Case Counts by Infection Date to Date of Report}
\usage{
adjust_infection_to_report(infections, delay_def, incubation_def)
}
\arguments{
\item{infections}{\code{data.table} containing a \code{date} variable and a numeric \code{cases} variable.}

\item{delay_def}{A single row data.table that defines the delay distribution (model, parameters and maximum delay for each model).
See \code{lognorm_dist_def} for an example of the structure.}

\item{incubation_def}{A single row data.table that defines the incubation distribution (model, parameters and maximum delay for each model).
See \code{lognorm_dist_def} for an example of the structure.}
}
\value{
A \code{data.table} containing a \code{date} variable (date of report) and a \code{cases} variable.
}
\description{
Stochastic mapping from cases by date of infection to date of report via date of
onset. Essentially reversal of \code{nowcast_pipeline}.
}
\examples{

## Define example cases
cases <- data.table::as.data.table(EpiSoon::example_obs_cases) 

cases <- cases[, `:=`(confirm = as.integer(cases), import_status = "local")]

## Define a single report delay distribution
delay_def <- EpiNow::lognorm_dist_def(mean = 5, 
                                      mean_sd = 1,
                                      sd = 3,
                                      sd_sd = 1,
                                      max_value = 30,
                                      samples = 1)
                                       
## Define a single incubation period
incubation_def <- EpiNow::lognorm_dist_def(mean = EpiNow::covid_incubation_period[1, ]$mean,
                                           mean_sd = EpiNow::covid_incubation_period[1, ]$mean_sd,
                                           sd = EpiNow::covid_incubation_period[1, ]$sd,
                                           sd_sd = EpiNow::covid_incubation_period[1, ]$sd_sd,
                                           max_value = 30, samples = 1)
                                           

## Perform a nowcast
nowcast <- nowcast_pipeline(reported_cases = cases, 
                            target_date = max(cases$date),
                            delay_defs = delay_def,
                            incubation_defs = incubation_def)
                            

infections <- nowcast[type \%in\% "infection_upscaled" & import_status \%in\% "local"]
infections <- infections[, `:=`(type = NULL, import_status = NULL)]

report <- adjust_infection_to_report(infections, delay_def, incubation_def)   

report                                   
}
