% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nowcasts.R
\name{clean_nowcasts}
\alias{clean_nowcasts}
\title{Clean Nowcasts for a Supplied Date}
\usage{
clean_nowcasts(date = NULL, nowcast_dir = NULL)
}
\arguments{
\item{date}{Date object. Defaults to todays date}

\item{nowcast_dir}{Character string giving the filepath to the nowcast results directory.}
}
\value{

}
\description{
USE WITH CARE. This function removes nowcasts in the format produced by \code{EpiNow} from a target
directory for the date supplied.
}
\examples{

}
