% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_report.R
\name{copy_report}
\alias{copy_report}
\title{Copy the Nowcast Report with an Appended YAML}
\usage{
copy_report(
  yaml = NULL,
  report = NULL,
  lines_to_cut = NULL,
  date = NULL,
  report_target = NULL
)
}
\arguments{
\item{yaml}{Character string giving the path to the YAML to append.}

\item{report}{Character string giving the path to the report.}

\item{lines_to_cut}{Numeric, number of lines to remove from the report. Defaults to \code{NULL}.
Useful when updating the author list etc.}

\item{date}{A character string date in the format \verb{"yyyy-mm-dd}. Indicates what date to add
to the report}

\item{report_target}{A character string identifying the target directory for the report.}
}
\value{

}
\description{
Copy the Nowcast Report with an Appended YAML
}
\examples{

}
