% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_dist_skel.R
\name{delay_dist_skel}
\alias{delay_dist_skel}
\title{Delay Distribution Skeleton}
\usage{
delay_dist_skel(n, dist = FALSE, cum = TRUE, model, params, max_delay = 120)
}
\arguments{
\item{n}{Numeric vector, number of samples to take (or days for the probability density).}

\item{dist}{Logical, defaults to \code{FALSE}. Should the probability density be returned rather
than a number of samples.}

\item{cum}{Logical, defaults to \code{TRUE}. If \code{dist = TRUE} should the returned distribution be
cumulative.}

\item{model}{Character string, defining the model to be used. Supported options are exponential
("exp") and gamma ("gamma").}

\item{params}{A list of parameters values (by name) required for each model. For the exponential model
this is a rate parameter and for the gamma model this is alpha and beta.}

\item{max_delay}{Numeric, the maximum delay to allow. Defaults to 120. Samples outside
of this range are resampled.}
}
\value{
A vector of samples or a probability distribution.
}
\description{
This function acts as a skeleton of for a truncated defined by
model type, maximum delay and model parameters. It is designed to be used with the
output from \code{get_delay_dist}.
}
\examples{

## Exponential model

## Sample
delay_dist_skel(10, model = "exp", params = list(rate = 1))

## Cumulative prob density
delay_dist_skel(1:10, model = "exp", dist = TRUE, params = list(rate = 1))

## Probability density
delay_dist_skel(1:10, model = "exp", dist = TRUE, 
                cum = FALSE, params = list(rate = 1))

## Gamma model

delay_dist_skel(10, model = "gamma", params = list(alpha = 1, beta = 2))

## Cumulative prob density
delay_dist_skel(0:10, model = "gamma", dist = TRUE,
                params = list(alpha = 1, beta = 2))

## Probability density
delay_dist_skel(0:10, model = "gamma", dist = TRUE, 
                cum = FALSE, params = list(alpha = 2, beta = 2))
    
}
