% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fit.R
\name{dist_fit}
\alias{dist_fit}
\title{Fit an integer adjusted exponential or gamma distribution}
\usage{
dist_fit(delays = NULL, samples = NULL, dist = "exp")
}
\arguments{
\item{delays}{Numeric vector of reporting delays}

\item{samples}{Numeric, number of samples to take}

\item{dist}{Character string, which distribution to fit. Defaults to exponential (\code{"exp"}) but
gamma is also supported (\code{"gamma"}).}
}
\value{

}
\description{
Fit an integer adjusted exponential or gamma distribution
}
\examples{

}
