% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eenb_fit.R
\name{eenb_fit}
\alias{eenb_fit}
\title{Fit the negative binomial version of EpiEstim}
\usage{
eenb_fit(nowcast_dir, geration_times)
}
\arguments{
\item{nowcast_dir}{The directory of the nowcast to fit to}
}
\value{

}
\description{
Fit the negative binomial version of EpiEstim
}
\examples{

}
