% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiNow-package.R
\docType{package}
\name{EpiNow-package}
\alias{EpiNow-package}
\alias{EpiNow}
\title{The 'EpiNow' package.}
\description{
To identify changes in the reproduction number, rate of spread,
and doubling time during the course of outbreaks whilst accounting
potential biases due to delays in case reporting.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
