% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_little_r.R
\name{estimate_little_r}
\alias{estimate_little_r}
\title{Estimate r}
\usage{
estimate_little_r(sample, min_time = NULL, max_time = NULL)
}
\arguments{
\item{sample}{A dataframe containing a numeric cases variable.}

\item{min_time}{Numeric, minimum time to use to fit the model.}

\item{max_time}{Numeric, maximum time to use to fit the model.}
}
\value{
A dataframe containing an estimate of r, its standard deviation and a
measure of the goodness of fit.
}
\description{
Estimate r
}
\examples{


}
