% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R0.R
\name{estimate_R0}
\alias{estimate_R0}
\title{Estimate the time varying R0 - using EpiEstim}
\usage{
estimate_R0(
  cases = NULL,
  serial_intervals = NULL,
  rt_prior = NULL,
  windows = NULL,
  si_samples = 100,
  rt_samples = 100,
  return_best = TRUE
)
}
\arguments{
\item{cases}{A dataframe containing a list of local cases with the following variables: \code{date}, \code{cases}, and \code{import_status}}

\item{serial_intervals}{A matrix with columns representing samples and rows representing the probability of the serial intervel being on
that day.}

\item{rt_prior}{A list defining the reproduction number prior containing the mean (\code{mean_prior}) and standard deviation (\code{std_prior})}

\item{windows}{Numeric vector, windows over which to estimate time-varying R. The best performing window will be
selected per serial interval sample by default (based on which window best forecasts current cases).}

\item{si_samples}{Numeric, the number of samples to take from the serial intervals supplied}

\item{rt_samples}{Numeric, the number of samples to take from the estimated R distribution for each time point.}

\item{return_best}{Logical defaults to \code{TRUE}. Should the estimates for the best performing window be returned or estimates
for all windows.}
}
\value{
A tibble containing the date and summarised R estimte.
}
\description{
Estimate the time varying R0 - using EpiEstim
}
\examples{

cases <- data.frame(date = seq(as.Date("2020-01-01"),
                               as.Date("2020-01-30"),
                               by = "days"),
                    cases = 1:30)
                    
                    
serial_intervals <- as.matrix(EpiNow::covid_serial_intervals[,1])
rt_prior <- list(mean_prior = 2.6, std_prior = 2) 
windows <- c(1, 3, 7)

estimate_R0(cases, serial_intervals, 
            rt_prior = rt_prior, windows = windows,
            rt_samples = 10, si_samples = 2, return_best = FALSE)
}
