% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_time_varying_measures_for_cases.R
\name{estimate_time_varying_measures_for_cases}
\alias{estimate_time_varying_measures_for_cases}
\title{Estimate time-varying measures for cases}
\usage{
estimate_time_varying_measures_for_cases(
  cases = NULL,
  serial_intervals = NULL,
  si_samples = NULL,
  rt_samples = NULL,
  start_rate_of_spread_est = NULL,
  window = NULL,
  rt_prior = NULL
)
}
\arguments{
\item{cases}{A dataframe containing a list of local cases with the following variables: \code{date}, \code{cases}, and \code{import_status}}

\item{serial_intervals}{A matrix with columns representing samples and rows representing the probability of the serial intervel being on
that day.}

\item{si_samples}{Numeric, the number of samples to take from the serial intervals supplied}

\item{rt_samples}{Numeric, the number of samples to take from the estimated R distribution for each time point.}

\item{start_rate_of_spread_est}{A character string in the form of a date ("2020-01-01")}

\item{window}{Numeric, the window over which to estimate time-varying R}

\item{rt_prior}{A list defining the reproduction number prior containing the mean (\code{mean_prior}) and standard deviation (\code{std_prior})}
}
\value{

}
\description{
Estimate time-varying measures for cases
}
\examples{


}
