% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_time_varying_r.R
\name{estimate_time_varying_r}
\alias{estimate_time_varying_r}
\title{Estimate time varying r}
\usage{
estimate_time_varying_r(onsets, window = 7)
}
\arguments{
\item{onsets}{A list of samples datasets nested within the dataset sampled from.}

\item{window}{}
}
\value{
A dataframe of r estimates over time summarisd across samples.
}
\description{
Estimate time varying r
}
\examples{

}
