% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_linelist.R
\name{generate_sample_linelist}
\alias{generate_sample_linelist}
\title{Generate a sample linelist from the observed linelist and sampled linelists}
\usage{
generate_sample_linelist(
  count_linelist = NULL,
  observed_linelist = NULL,
  merge_actual_onsets = TRUE,
  earliest_allowed_onset = NULL
)
}
\arguments{
\item{count_linelist}{Dataframe with two variables: date_report and daily_linelist. As generated by  \code{linelist_from_case_counts}.}

\item{observed_linelist}{Dataframe with two variables: date_report and daily_observed_linelist. As generated by `split_linelist_by_day``}

\item{merge_actual_onsets}{Logical, defaults to \code{TRUE}. Should linelist onset dates be used where available?}

\item{earliest_allowed_onset}{A character string in the form of a date ("2020-01-01") indiciating the earliest
allowed onset.}
}
\value{
Dataframe with no missing data and two variables: date_report and date_onset
}
\description{
Generate a sample linelist from the observed linelist and sampled linelists
}
\examples{


}
