% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delay_sample_fn.R
\name{get_delay_dist}
\alias{get_delay_dist}
\title{Get a Parameters that Defne a Discrete Delay}
\usage{
get_delay_dist(
  delays,
  verbose = FALSE,
  samples = 1,
  bootstraps = 1,
  bootstrap_samples = 250
)
}
\arguments{
\item{delays}{Numeric vector of integer delays.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should progress messages be printed}

\item{bootstraps}{Numeric, defaults to 1. The number of bootstrap samples (with replacement)
of the delay distribution to take.}

\item{bootstrap_samples}{Numeric, defaults to 100. The number of samples to take in each boostrap.
When the sample size of the supplied delay distribution is less than 100 this is used instead.}
}
\value{
A data.table of delay distributions and the parameters that define them.
}
\description{
Get a Parameters that Defne a Discrete Delay
}
\examples{

## Example with exponential and a small smaple
delays <- rexp(20, 1)

get_delay_dist(delays, samples = 10, verbose = TRUE)


## Example with gamma and a larger sample
delays <- rgamma(50, 1, 2)

get_delay_dist(delays, sample = 10, bootstraps = 2)


}
\author{
Sebastian Funk \href{mailto:sebastian.funk@lshtm.ac.uk}{sebastian.funk@lshtm.ac.uk}
}
