% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delay_sample_fn.R
\name{get_delay_sample_fn}
\alias{get_delay_sample_fn}
\title{Get a sampling function for reporting delays}
\usage{
get_delay_sample_fn(linelist, verbose = FALSE, samples = 1)
}
\arguments{
\item{linelist}{Dataframe with a delay_confirmation date variable}

\item{verbose}{Logical, defaults to \code{FALSE}. Should progress messages be printed}

\item{resample}{Logical, defaults to \code{FALSE}. Should the supplied delays be resampled.}
}
\value{
A list of function that takes one parameter, \code{n}, the number of reporting
delays to randomly sample
}
\description{
Get a sampling function for reporting delays
}
\examples{

## Example
\dontrun{
get_delay_sample_fn()
}

## Code
get_delay_sample_fn
}
\author{
Sebastian Funk \href{mailto:sebastian.funk@lshtm.ac.uk}{sebastian.funk@lshtm.ac.uk}
}
