% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Get Timeseries from EpiNow}
\usage{
get_timeseries(results_dir = NULL, date = NULL)
}
\arguments{
\item{results_dir}{A character string indicating the folder containing the \code{EpiNow}
results to extract.}

\item{date}{A Character string (in the format "yyyy-mm-dd") indicating the date to extract
data for}
}
\value{

}
\description{
Get Timeseries from EpiNow
}
\examples{

## Code
get_timeseries
}
