% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_map.R
\name{global_map}
\alias{global_map}
\title{Generate a global map for a single variable.}
\usage{
global_map(
  data = NULL,
  variable = NULL,
  variable_label = NULL,
  trans = "identity",
  fill_labels = NULL,
  viridis_palette = "cividis",
  show_caption = TRUE
)
}
\arguments{
\item{data}{Dataframe containing variables to be mapped. Must contain a \code{country} variable.}

\item{variable}{A character string indicating the variable to map data for. This must be supplied.}

\item{variable_label}{A character string indicating the variable label to use. If not supplied then the underlying
variable name is used.}

\item{trans}{A character string specifying the transform to use on the specified metric. Defaults to no
transform ("identity"). Other options include log scaling ("log") and log base 10 scaling
("log10"). For a complete list of options see \code{ggplot2::continous_scale}.}

\item{fill_labels}{A function to use to allocate legend labels. An example (used below) is \code{scales::percent},
which can be used for percentage data.}

\item{viridis_palette}{Character string indicating the \code{viridis} colour palette to use. Defaults
to "cividis". Options include "cividis", "magma", "inferno", "plasma", and "viridis". For additional details}
}
\value{
A \code{ggplot2} object containing a global map.
}
\description{
This general purpose function can be used to generate a global map for a single variable. It has few defaults but
the data supplied must contain a \code{country} variable for linking to mapping data.
}
\examples{


}
