% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report_delay.R
\name{plot_report_delay}
\alias{plot_report_delay}
\title{Plot report delay}
\usage{
plot_report_delay(linelist, cutoff, drop_from_plot)
}
\arguments{
\item{cutoff}{Date from when data will be used in the analysis.}

\item{drop_from_plot}{Date from which to display data.}

\item{time_delays}{A linelist containing \code{date_confirm} and \code{report_delay} columns}
}
\value{
A ggplot2 object
}
\description{
Plot report delay
}
\examples{


}
