% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{Plot a Summary of the Latest Results}
\usage{
plot_summary(summary_results, x_lab = "Region")
}
\arguments{
\item{summary_results}{A dataframe as returned by \code{summarise_results} (the \code{data} object).}

\item{x_lab}{A character string giving the label for the x axis, defaults to region.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot a Summary of the Latest Results
}
\examples{


}
