% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{R_to_r}
\alias{R_to_r}
\title{Convert Reproduction Numbers to Growth Rates}
\usage{
R_to_r(R, gamma_mean, gamma_sd)
}
\arguments{
\item{R}{Numeric, Reproduction number estimates}

\item{gamma_sd}{Numeric, standard deviation of the gamma distribution}
}
\value{
Numeric vector of reproduction number estimates
}
\description{
Convert Reproduction Numbers to Growth Rates
}
\examples{

R_to_r(2, 4, 1)  
}
