% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbinom_size.R
\name{rbinom_size}
\alias{rbinom_size}
\title{Draw with an offset from a negative binomial distribution}
\usage{
rbinom_size(n, x, prob)
}
\arguments{
\item{n}{Numeric, number of samples to draw}

\item{x}{Numeric, offset.}

\item{prob}{Numeric, probability of successful trial}
}
\description{
Samples size (the number of trials) of a binomial distribution
copied from https://github.com/sbfnk/bpmodels/blob/master/R/utils.r
}
