% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_summary.R
\name{regional_summary}
\alias{regional_summary}
\title{Generate Regional Summary Output}
\usage{
regional_summary(
  results_dir = NULL,
  summary_dir = NULL,
  target_date = NULL,
  region_scale = "Region"
)
}
\arguments{
\item{summary_dir}{A character string giving the directory
in which to store summary of results.}

\item{target_date}{A character string giving the target date for which to extract results
(in the format "yyyy-mm-dd").}

\item{region_scale}{A character string giving the label to use for regions.}
}
\description{
Generate Regional Summary Output
}
\examples{

\dontrun{

## Example asssumes that CovidGlobalNow (github.com/cmmid/CovidGlobalNow) is  
## in the directory above the root.
regional_summary(results_dir = "../CovidGlobalNow/results",
                 summary_dir = "../CovidGlobalNow/summary",
                 target_date = "2020-03-19",
                 region_scale = "Country/Region")

## Example asssumes that CovidItalyNow (github.com/cmmid/CovidItalyNow) is  
## in the directory above the root.
regional_summary(results_dir = "../CovidItalyNow/results",
                 summary_dir = "../CovidItalyNow/summary",
                 target_date = "2020-03-19",
                 region_scale = "Region")
}

}
