% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report_summary}
\alias{report_summary}
\title{Provide Summary Statistics on an Rt Pipeline}
\usage{
report_summary(target_folder)
}
\value{

}
\description{
Provide Summary Statistics on an Rt Pipeline
}
\examples{

}
