% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_approx_delay.R
\name{sample_approx_delay}
\alias{sample_approx_delay}
\title{Approximate Sampling Reporting Delays using Reported Case Counts}
\usage{
sample_approx_delay(
  reported_cases = NULL,
  delay_fn = NULL,
  max_delay = 120,
  earliest_allowed_onset = NULL,
  direction = "onset"
)
}
\arguments{
\item{reported_cases}{A dataframe of reported cases (in date order) with the following variables:
\code{date} and \code{confirm}.}

\item{delay_fn}{A sampling funtion that takes a single numeric argument and returns a vector of
numeric samples this long.}

\item{max_delay}{Numeric, maximum delay to allow. Defaults to 120 days}

\item{earliest_allowed_onset}{A character string in the form of a date ("2020-01-01") indiciating the earliest
allowed onset.}

\item{direction}{Character string, defato "onset". Direction in which to map cases. Supports
either from report to onset ("onset") or from onset to report ("report")}
}
\value{
A \code{data.table} of cases by date of onset
}
\description{
Approximate Sampling Reporting Delays using Reported Case Counts
}
\examples{

cases <- data.table::as.data.table(EpiSoon::example_obs_cases) 

cases <- cases[, confirm := cases] 

delay_fn <- function(n, dist, cum) {
   dgamma(n, 2, 1)}

onsets <- sample_approx_delay(reported_cases = cases,
                              delay_fn = delay_fn)
                              
reports <- sample_approx_delay(reported_cases = cases,
                              delay_fn = delay_fn,
                              direction = "report")

}
