% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{sample_approx_dist}
\alias{sample_approx_dist}
\title{Approximate Sampling a Distribution using Counts}
\usage{
sample_approx_dist(
  reported_cases = NULL,
  dist_fn = NULL,
  max_value = 120,
  earliest_allowed_onset = NULL,
  direction = "onset"
)
}
\arguments{
\item{reported_cases}{A dataframe of reported cases (in date order) with the following variables:
\code{date} and \code{confirm}.}

\item{earliest_allowed_onset}{A character string in the form of a date ("2020-01-01") indiciating the earliest
allowed onset.}

\item{direction}{Character string, defato "onset". Direction in which to map cases. Supports
either from report to onset ("onset") or from onset to report ("report")}

\item{max_delay}{Numeric, maximum delay to allow. Defaults to 120 days}
}
\value{
A \code{data.table} of cases by date of onset
}
\description{
Approximate Sampling a Distribution using Counts
}
\examples{

cases <- data.table::as.data.table(EpiSoon::example_obs_cases) 

cases <- cases[, confirm := as.integer(cases)] 

## Reported case distribution
print(cases)

## Total cases
sum(cases$confirm)

delay_fn <- function(n, dist, cum) {
   pgamma(n + 0.9999, 2, 1) - pgamma(n - 1e-5, 2, 1)}

onsets <- sample_approx_dist(reported_cases = cases,
                             dist_fn = delay_fn)
   
## Estimated onset distribution
print(onsets)
  
## Check that sum is equal to reported cases
total_onsets <- median(
   purrr::map_dbl(1:100, 
                  ~ sum(sample_approx_dist(reported_cases = cases,
                  dist_fn = delay_fn)$cases))) 
                   
total_onsets
 
                   
## Map from onset cases to reported                  
reports <- sample_approx_dist(reported_cases = cases,
                              dist_fn = delay_fn,
                              direction = "report")

}
