% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_onsets.R
\name{sample_onsets}
\alias{sample_onsets}
\title{Sample onset times}
\usage{
sample_onsets(
  onsets = NULL,
  cum_freq = NULL,
  dates = NULL,
  samples = NULL,
  report_delay = NULL
)
}
\arguments{
\item{onsets}{Numeric vector of onset times}

\item{cum_freq}{Numeric vector of cumulative frequencies}

\item{dates}{Character vector of dates}

\item{samples}{Numeric, number of samples to take}

\item{report_delay}{Numeric, delay between the last data point and the date of the report}
}
\value{

}
\description{
Sample onset times
}
\examples{


}
