% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cases}
\alias{simulate_cases}
\title{Simulate Cases by Date of Infection, Onset and Report}
\usage{
simulate_cases(
  rts,
  initial_cases,
  initial_date,
  generation_interval,
  rdist = rpois,
  delay_def,
  incubation_def,
  reporting_effect,
  type = "sample"
)
}
\arguments{
\item{rts}{A dataframe of containing two variables \code{rt} and  \code{date} with
\code{rt} being numeric and \code{date} being a date.}

\item{initial_cases}{Integer, initial number of cases.}

\item{initial_date}{Date, (i.e \code{as.Date("2020-02-01")}). Starting date of the simulation.}

\item{generation_interval}{Numeric vector describing the generation interval probability density}

\item{rdist}{A function to be used to sample the number of cases. Must take two
arguments with the first specfying the number of samples and the second the mean. Defaults
to \code{rpois} if not supplied}

\item{delay_def}{A single row data.table that defines the delay distribution (model, parameters and maximum delay for each model).
See \code{lognorm_dist_def} for an example of the structure.}

\item{incubation_def}{A single row data.table that defines the incubation distribution (model, parameters and maximum delay for each model).
See \code{lognorm_dist_def} for an example of the structure.}

\item{reporting_effect}{A numeric vector of length 7 that allows the scaling of reported cases
by the day on which they report (1 = Monday, 7 = Sunday). By default no scaling occurs.}

\item{type}{Character string indicating the method to use to transfrom counts. Supports either "sample"
which approximates sampling or "median" would shift by the median of the distribution.}
}
\value{
A dataframe containing three variables: \code{date}, \code{cases} and \code{reference}.
}
\description{
Simulate Cases by Date of Infection, Onset and Report
}
\examples{
## Define an initial rt vector 
rts <- c(rep(2, 20), (2 - 1:15 * 0.1), rep(0.5, 10))
rts
## Use the mean default generation interval for covid
generation_interval <- rowMeans(EpiNow::covid_generation_times)

## Sample a report delay as a lognormal
delay_def <- EpiNow::lognorm_dist_def(mean = 5, mean_sd = 1,
                                      sd = 3, sd_sd = 1, max_value = 30,
                                       samples = 1)
                                      

## Sample a incubation period (again using the default for covid)
incubation_def <- EpiNow::lognorm_dist_def(mean = EpiNow::covid_incubation_period[1, ]$mean,
                                          mean_sd = EpiNow::covid_incubation_period[1, ]$mean_sd,
                                          sd = EpiNow::covid_incubation_period[1, ]$sd,
                                          sd_sd = EpiNow::covid_incubation_period[1, ]$sd_sd,
                                          max_value = 30, samples = 1)

## Simulate cases with a decrease in reporting at weekends and an incease on Monday                                     
simulated_cases <- simulate_cases(rts, initial_cases = 100 , initial_date = as.Date("2020-03-01"),
                    generation_interval = generation_interval, delay_def = delay_def, 
                   incubation_def = incubation_def, reporting_effect = c(1.1, rep(1, 4), 0.95, 0.95))
                   
print(simulated_cases)
}
