% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist_tools.R
\name{split_linelist_by_day}
\alias{split_linelist_by_day}
\title{Convert a linelist into a nested tibble of linelists by day}
\usage{
split_linelist_by_day(linelist = NULL)
}
\arguments{
\item{linelist}{Dataframe with the following variables date_onset_symptoms and date_confirmation}
}
\value{
A nested tibble with a linelist per day (daily_observed_linelist) variable containing date_onset and date_report and
a date_report variable
}
\description{
Convert a linelist into a nested tibble of linelists by day
}
\examples{


}
