% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_as_csv.R
\name{summarise_to_csv}
\alias{summarise_to_csv}
\title{Summarise rt and cases as a csv}
\usage{
summarise_to_csv(
  results_dir = NULL,
  summary_dir = NULL,
  type = "country",
  incubation_shift = 5
)
}
\arguments{
\item{results_dir}{Character string indicating the directory from which to extract results}

\item{summary_dir}{Character string the directory into which to save results}

\item{type}{Character string, the region identifier to apply}

\item{incubation_shift}{Numeric, the incubation shift to apply to the estimates, defaults to 5 days.}
}
\value{
Nothing is returned
}
\description{
Summarise rt and cases as a csv
}
\examples{


}
