% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{Custom Map Theme}
\usage{
theme_map(map = NULL, continuous = FALSE)
}
\arguments{
\item{map}{\code{ggplot2} map object}

\item{continuous}{Logical defaults to \code{FALSE}. Is the fill variable continuous.}
}
\value{
A \code{ggplot2} object
}
\description{
Custom Map Theme
}
\examples{


## Code 
theme_map
}
