% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_R0.R
\name{estimate_R0}
\alias{estimate_R0}
\title{Estimate the time varying R0 - using EpiEstim}
\usage{
estimate_R0(
  cases = NULL,
  serial_intervals = NULL,
  rt_prior = NULL,
  window = NULL,
  si_samples = 100,
  rt_samples = 100
)
}
\arguments{
\item{cases}{A dataframe containing a list of local cases with the following variables: \code{date}, \code{cases}, and \code{import_status}}

\item{serial_intervals}{A matrix with columns representing samples and rows representing the probability of the serial intervel being on
that day.}

\item{rt_prior}{A list defining the reproduction number prior containing the mean (\code{mean_prior}) and standard deviation (\code{std_prior})}

\item{window}{Numeric, the window over which to estimate time-varying R}

\item{si_samples}{Numeric, the number of samples to take from the serial intervals supplied}

\item{rt_samples}{Numeric, the number of samples to take from the estimated R distribution for each time point.}
}
\value{
A tibble containing the date and summarised R estimte.
}
\description{
Estimate the time varying R0 - using EpiEstim
}
\examples{


}
