% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_max_var.R
\name{pull_max_var}
\alias{pull_max_var}
\title{Pull out maximum variables values based on a variable}
\usage{
pull_max_var(df, max_var = NULL, sel_var = NULL, type_selected = NULL)
}
\arguments{
\item{df}{Datatable with the following variables: \code{type} and \code{si_dist}}

\item{type_selected}{The nowcast type to extract.}

\item{var}{Unquoted variable name to pull out the maximum R estimate for.}
}
\value{
A character string containing the maximum variable
}
\description{
Pull out maximum variables values based on a variable
}
\examples{

df <- data.table::data.table(type = c("nowcast", "other"),
                             var = c(1:10),
                             sel = "test")
                             
pull_max_var(df, max_var = "var", sel_var = "var", type_selected = "nowcast")
}
