% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_pipeline.R
\name{rt_pipeline}
\alias{rt_pipeline}
\title{Real-time Pipeline}
\usage{
rt_pipeline(
  cases = NULL,
  imported_cases = NULL,
  linelist = NULL,
  target_folder = NULL,
  target_date = NULL,
  delay_cutoff_date = NULL,
  predict_lag = 0,
  samples = 1000,
  si_samples = 1,
  rt_samples = 100,
  window = 7,
  earliest_allowed_onset = NULL,
  start_rate_of_spread_est = NULL,
  merge_actual_onsets = TRUE,
  case_only = FALSE,
  delay_only = FALSE,
  verbose = FALSE,
  serial_intervals = NULL,
  rt_prior = NULL,
  save_plots = FALSE
)
}
\arguments{
\item{cases}{A dataframe of cases (\code{confirm}) by date of confirmation (\code{date}) and import status (\code{import_status})}

\item{linelist}{A dataframe of of cases (by row) containing the following variables:
\code{import_status} (values "local" and "imported"), \code{date_onset}, \code{date_confirm}, \code{report_delay}.}

\item{target_folder}{Character string, name of the folder in which to save the results.}

\item{target_date}{Character string, in the form "2020-01-01". Date to cast.}

\item{delay_cutoff_date}{Character string, in the form "2020-01-01". Cutoff date to use
to estimate the delay distribution.}

\item{predict_lag}{Numeric, number of days to use as a prediction lag. Defaults to 0.}

\item{samples}{Numeric, the number of pseudo linelists to generate. Defaults to 1000.}

\item{si_samples}{Numeric, the number of samples to take from the serial intervals supplied}

\item{rt_samples}{Numeric, the number of samples to take from the estimated R distribution for each time point.}

\item{window}{Numeric, the window to use to estimate time-varying parameters. Defaults to 7.}

\item{earliest_allowed_onset}{A character string in the form of a date ("2020-01-01") indiciating the earliest
allowed onset.}

\item{start_rate_of_spread_est}{A character string in the form of a date ("2020-01-01") indicating the start time to
estimate the rate of sprea.}

\item{case_only}{Logical, defaults to \code{FALSE}. Should estimates also be made based on reported cases.}

\item{verbose}{Logical, defaults to \code{FALSE}. Should internal nowcasting progress messages be returned.}

\item{serial_intervals}{A matrix with columns representing samples and rows representing the probability of the serial intervel being on
that day. Defaults to \code{EpiNow::covid_serial_intervals}.}

\item{rt_prior}{A list defining the reproduction number prior containing the mean (\code{mean_prior}) and standard deviation (\code{std_prior})}

\item{save_plots}{Logical, defaults to code \code{FALSE}. Should plots be saved.}
}
\description{
Combine fitting a delay distribution, constructing a set of
complete sampled linelists, nowcast cases by onset date, and estimate
the time-varying effective reproduction number and rate of spread.
}
\examples{

}
